requirements.txt
altdss==0.2.4
cffi==1.17.1
contourpy==1.3.1
cycler==0.12.1
dss-python==0.15.7
dss_python_backend==0.14.5
fonttools==4.55.3
kiwisolver==1.4.8
matplotlib==3.10.0
networkx==3.4.2
numpy==2.2.1
opendssdirect.py==0.9.4
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pycparser==2.22
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2024.2
scipy==1.14.1
six==1.17.0
typing_extensions==4.12.2
tzdata==2024.2
